[SPARK-9242][SQL] Audit UDAF interface.
